while loop

+++ CS01-13.py
@@ -1,4 +1,4 @@
-i = int(input('ใส่คะแนน : '))
+i = float(input('ใส่คะแนน : '))
 while i != -1:
     print(i)
     i = int(input('ใส่คะแนน : '))
